fix: refactor sample_donor
