+++ test/unit/model/test_sample_donor.py
@@ -1,7 +1,7 @@
 import unittest
 from datetime import datetime
 
-from MIABIS_on_FHIR.MoF_sample_donor import SampleDonor
+from MIABIS_on_FHIR.sample_donor import SampleDonor
 from MIABIS_on_FHIR.gender import Gender
